Add a vect_worthwhile_without_simd_p helper routine
For some operations we'll consider lowering "vector" operations
into N scalar word operations.  This N needs to be fixed at
compile time, so the condition guarding it needs to change
when variable-lengh vectors are added.  This patch puts the
condition into a helper routine so that there's only one
place to update.