Use MACRO_MODE for CANNOT_CHANGE_MODE_CLASS
This patch uses MACRO_MODE in calls to CANNOT_CHANGE_MODE_CLASS.
The macro is still used conditionally, so it's not trivial to
convert it to a target hook.  It would probably also be a good
idea to redesign this interface when converting it to a hook,
or at least change the name so that it doesn't lead to double
negatives.