Allow the number of iterations to be smaller than VF
Fully-masked loops can be profitable even if the iteration
count is smaller than the vectorisation factor.  In this case
we're effectively doing a complete unroll followed by SLP.

The documentation for min-vect-loop-bound says that the
default value was 0, but actually the default and minimum
were 1.  We need it to be 0 for this case since the parameter
counts a whole number of vector iterations.

The patch also fixes a case in which the threshold could
wrap around to a very large unsigned number, which is why
we didn't use SVE and ST2 for f1 in sve_cant_predicate_1.c.